Changed the position of authPass from controller to middleware
