+++ back-end/routers/auth.js
@@ -1,7 +1,7 @@
 import express from 'express';
 const router = express.Router();
 
-import { register, login } from '../controllers/authController.js';
+import { register, login } from '../controllers/auth.js';
 
 router.route('/register').post(register);
 router.route('/login').post(login);
